Create README.md
Designing adaptor molecules against protein targets is a challenging task. In recent years, deep generative models have been successfully applied to the molecular design of new drugs. Models that construct molecules from the atomic level are prone to generate structures that do not conform to the chemical laws; the development of molecular design models that are easy to synthesize and have high affinity remains an important part of research in this field. In this paper, we develop a target protein-oriented drug molecular design model, BFMOL. The model consists of two encoders and a gated self-attentive decoder; the encoder is used to extract the features of the target and ligand molecules, and the decoder realizes the association of the two features and decodes them into the probability of the molecular fragments, which then generates the ligand molecules. The model fragments the molecule by BRICS method and characterizes the target at three levels: atom, residue and inter-residue. The results show that on the Crossdocked dataset, the molecules designed by BFMOL have an average synthetic accessibility score of 0.769 and an average druggability score of 0.607, which are higher than the baseline model PMDM (0.611 and 0.594). In the design of molecules against kinase-like targets, the model-generated molecules possessed more double six-membered ring systems with typical features of kinase inhibitors. In substructure analysis, the model-generated molecules have the closest ring structure and bond angle distribution to the real molecules. In conclusion, a more rational neural network model for potential drug molecule design toward target proteins was developed in this work